requirements.txt
openai
streamlit
python-dotenv
azure-cosmos
tiktoken